fix vline in exported STACK diagplot
The contour plot for the stacking of error surfaces contains horizontal and vertical blue lines indicating the best solution after stacking. However, in the exported diagnostic plot (both png and pdf format) for all three error surface stacking options (no weight, WS, RH) the vertical line is missing and only the horizontal line is visible. Only the last plotted line of the two lines is affected by this issue. Here this is simply fixed by plotting the vertical line twice. For the sake of safety also the horizontal line is plot twice.